requirements.txt
python-docx>=0.8.11
Pillow>=9.0.0
streamlit>=1.28.0
requests>=2.31.0
